requirements.txt
setuptools==67.8.0
camelot-py[cv]==0.11.0
PyPDF2==2.12.1
pymupdf==1.22.3
dpath==2.1.6